requirements.txt
absl-py==0.7.0
aioconsole==0.1.7
amqp==2.4.2
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.3.0
attrs==17.4.0
Automat==0.6.0
backports.functools-lru-cache==1.5
billiard==3.6.0.0
blis==0.2.4
celery==4.3.0
celery-progress==0.0.3
certifi==2018.1.18
cffi==1.12.2
chardet==3.0.4
click==6.7
cloudpickle==0.8.1
colorama==0.4.1
configobj==5.0.6
constantly==15.1.0
coverage==4.4.2
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
Cython==0.29.7
dask==1.2.0
DateTime==4.3
decorator==4.4.0
Django==2.1.7
django-cleanup==3.2.0
easygui==0.98.1
enum34==1.1.6
funcsigs==1.0.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.8
incremental==16.10.1
ipaddress==1.0.22
jsonschema==2.6.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
kombu==4.5.0
Markdown==3.0.1
matplotlib==3.0.3
mock==2.0.0
more-itertools==5.0.0
mpl-finance==0.10.0
murmurhash==1.0.2
numpy==1.16.2
opencv-python==4.1.0.25
pandas==0.24.2
pbr==3.1.1
Pillow==6.0.0
pluggy==0.9.0
protobuf==3.7.1
py==1.8.0
pycparser==2.19
pyparsing==2.4.0
pytest==4.4.1
python-dateutil==2.8.0
pytz==2018.9
pywin32-ctypes==0.2.0
PyYAML==3.12
scipy==1.2.1
SecretStorage==2.3.1
selenium==3.141.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.2
vine==1.3.0
Werkzeug==0.15.2
zope.interface==4.6.0
